Follow up 6545fde8f16024bfa5097fab95637764f0b6b757
Switched all EXP fields over to long. Seems like some were already using the correct type, we weren't very consistent there^^"

+++ src/Shared/Network/Helpers/CommanderHelper.cs
@@ -42,8 +42,8 @@ namespace Melia.Shared.Network.Helpers
 	{
 		int Handle { get; }
 		Position Position { get; }
-		int Exp { get; }
-		int MaxExp { get; }
+		long Exp { get; }
+		long MaxExp { get; }
 		long TotalExp { get; }
 		long Id { get; }
 		int Hp { get; }
